use interactive shell to execute .fpp.sh
This conveniently executes all the appropriate rc files for both zsh and
bash, thus expanding out aliases. There are cons to this approach
though:

* Though the Bash documentation strongly suggests that expand_aliases is
  turned on when interactive, I was not able to reproduce this behavior
  on Yosemite bash. (It's possible Yosemite just comes with a weird
  setup though.)

* $SHELL only ever expands out to the user's default shell and not the
  current shell. (So in the rare case where you might open up zsh, then
  run bash, and then use fpp, fpp will choose zsh and not bash.)

* Not all shells support the -i flag. However, the two shells currently
  supported by fpp (zsh and bash) do indeed have this handy dandy flag.

A controversial change, perhaps! My feelings would not be hurt if you
guys chose to not accept this PR, but my heart would be ecstatic if you
did.

This fixes issue #125.