Change path of output file to match debian standard
The standard debian filename of a package is
PACKAGENAME_VERSION_ARCH.deb. Therefore, ideally it should be named
pathpicker_0.9.2_all.deb. In the Debian manual at
https://www.debian.org/doc/manuals/debian-faq/pkg-basics.en.html#pkgname,
it says:

    The Debian binary package file names conform to the following
    convention: <foo>_<VersionNumber>-<DebianRevisionNumber>_<DebianArchitecture>.deb

This currently adds an extra step before we can add it to our local repo
so it would be nice if this output the expected filename.